requirements.txt
colorama>=0.4.6
inquirer>=3.1.2
PyYAML>=6.0
